Use MUnit in UriConvertibleSuite
